Tweaked OpenGL context construction
The constructors now return the context if everything is valid.
If not valid it returns None.
The OpenGL 2.1 constructor now checks if GL_ARB_explicit_attrib_location is supported by the system.
If the system does not support it an error will be raised at construction rather than when drawing.